Fix: warnings in SofaFramework
Former-commit-id: d90d3717b4622a3923ea235dbe1440668ae12209

+++ framework/sofa/core/objectmodel/Base.cpp
@@ -474,7 +474,7 @@ bool Base::parseField( const std::string& attribute, const std::string& value)
             ok = false;
         }
         sout << "Link " << linkVec[l]->getName() << " = " << linkVec[l]->getValueString() << sendl;
-        unsigned int s = linkVec[l]->getSize();
+        size_t s = linkVec[l]->getSize();
         for (unsigned int i=0; i<s; ++i)
         {
             sout  << "  " << linkVec[l]->getLinkedPath(i) << " = ";
